refactor: Rename permissionError file to permissionUtils, move hasPermission to permissionUtils.
