+++ src/classes/AdvancedClient.ts
@@ -32,20 +32,4 @@ export class AdvancedClient extends Client {
 	get username() {
 		return this.user?.username;
 	}
-
-	/**
-	 * Tells you if the client has permissions from a message in a fancy way than the {@link https://discord.js.org/#/docs/main/stable/class/GuildMember?scrollTo=hasPermission | GuildMember#hasPermission} method.
-	 *
-	 * @param message - The message to check permissions from.
-	 * @param permission - The permission to check.
-	 * @returns If the user has the permission.
-	 */
-	public hasPermission(message: Message, permission: PermissionResolvable) {
-		return message.guild
-			? message.guild.me?.hasPermission(permission, {
-					checkOwner: false,
-					checkAdmin: false,
-			  })
-			: false;
-	}
 }
