+++ src/classes/Command.ts
@@ -1,6 +1,6 @@
 import {DMChannel, GuildChannel, GuildMember, Message, PermissionOverwrites, Permissions, PermissionString, Snowflake, TextChannel, User} from 'discord.js';
 import {CommandHandler} from '../CommandHandler';
-import {isPermission} from '../utils/permissionsError.js';
+import {isPermission} from '../utils/permissionUtils.js';
 import {isOwner} from '../utils/utils';
 import {CommandContext} from './CommandContext.js';
 import CommandCooldown = CommandHandler.CommandCooldown;
