+++ src/utils/permissionUtils.ts
@@ -1,4 +1,4 @@
-import {Message, Permissions, PermissionString} from 'discord.js';
+import {Message, PermissionResolvable, Permissions, PermissionString} from 'discord.js';
 import {BetterEmbed} from 'discord.js-better-embed';
 import {Command} from '../classes/Command';
 
@@ -31,4 +31,20 @@ export function permissionsError(message: Message, missingPermissions: Permissio
  */
 export function isPermission(permission: string): permission is PermissionString {
 	return Object.keys(Permissions.FLAGS).includes(permission);
+}
+
+/**
+ * Tells you if the client has permissions from a message in a fancy way than the {@link https://discord.js.org/#/docs/main/stable/class/GuildMember?scrollTo=hasPermission | GuildMember#hasPermission} method.
+ *
+ * @param message - The message to check permissions from.
+ * @param permission - The permission to check.
+ * @returns If the user has the permission.
+ */
+function hasPermission(message: Message, permission: PermissionResolvable) {
+	return message.guild
+		? message.guild.me?.hasPermission(permission, {
+				checkOwner: false,
+				checkAdmin: false,
+		  })
+		: false;
 }
